Drop Java 6 support
We dropped Java 6 support in `geoip2` over a year ago and at the time,
agreed to do it here as well. We never got around to it here, likely due
to the slower release cycle of this library.

I am motivated to do it now as Travis does not support Java 6 on its
Trusty images.

Also, dropping testing of `oraclejdk7` as that is also no longer
available on Trusty images.