Pass Type parameter through when loading via cache
Previously the lack of this meant that we'd not decode to the expected
type in cases such as List<Foo>. This was because Type wasn't passed
through given the call to decode() didn't include it.

This also now takes into account the class and type in the cache key,
where previously we could behave return a different class than expected
if deserializing into two different classes.