Specify java release to use when building
This is because we see

java.lang.NoSuchMethodError:
java.nio.ByteBuffer.position(I)Ljava/nio/ByteBuffer;

when running on Java 8 when using this in GeoIP2-java, as it gets built
with a newer version of Java apparently.

See https://github.com/eclipse/jetty.project/issues/3244.

Do this with a profile because Java 8 does not support the release flag,
which is required for this.