Remove ossrh-snapshots repository
Dependabot will stop looking for maven dependencies once it finds
a repository that returns any number of results, even if none of
those results match the requested version. This removes the ossrh
repository because it brings in snapshot repositories that prevent
dependabot from falling back to the central repository. See:
dependabot/dependabot-core#5872